requirements.txt
chat_downloader
scrapetube
flask
humanize
flask[async] 
psycopg2-binary